Remove self_registration models, reusing existing self_registration controller
