+++ app/views/groups/self_registration/_form.html.haml
@@ -1,7 +1,6 @@
--#  Copyright (c) 2023-2024, Schweizer Alpen-Club. This file is part of
--#  hitobito_sac_cas and licensed under the Affero General Public License version 3
--#  or later. See the COPYING file at the top-level directory or at
--#  https://github.com/hitobito/hitobito
+-#  Copyright (c) 2024, Schweizer Alpen-Club. This file is part of hitobito and licensed under the
+-#  Affero General Public License version 3 or later. See the COPYING file at the top-level directory
+-#  or at https://github.com/hitobito/hitobito.
 
-= standard_form(entry, url: group_self_registration_path, data: { turbo: 'false' }) do |f|
-  = render 'fields', f: f
+= standard_form(wizard, url: group_self_registration_path, data: { controller: 'forwarder'}) do |f|
+  = render StepsComponent.new(partials: wizard.partials, step: wizard.current_step, form: f)
