+++ app/views/groups/self_registration/show.html.haml
