--- app/views/wizards/register_new_users/_form.html.haml
@@ -1,6 +1,0 @@
--#  Copyright (c) 2024, Schweizer Alpen-Club. This file is part of hitobito and licensed under the
--#  Affero General Public License version 3 or later. See the COPYING file at the top-level directory
--#  or at https://github.com/hitobito/hitobito.
-
-= standard_form(wizard, url: group_register_new_user_path, data: { controller: 'forwarder'}) do |f|
-  = render StepsComponent.new(partials: wizard.partials, step: wizard.current_step, form: f)
