+++ spec/controllers/groups/self_registration_controller_spec.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
-#  Copyright (c) 2021, Efficiency-Club Bern. This file is part of
-#  hitobito and licensed under the Affero General Public License version 3
-#  or later. See the COPYING file at the top-level directory or at
-#  https://github.com/hitobito/hitobito.
+#  Copyright (c) 2024, Schweizer Alpen-Club. This file is part of hitobito and licensed under the
+#  Affero General Public License version 3 or later. See the COPYING file at the top-level directory
+#  or at https://github.com/hitobito/hitobito.
 
 require "spec_helper"
 
 describe Groups::SelfRegistrationController do
+  let(:role_type) { Group::TopGroup::Member }
   let(:group) { groups(:top_group) }
   let(:person) { people(:top_leader) }
 
@@ -17,28 +17,27 @@ describe Groups::SelfRegistrationController do
       allow(Settings.groups.self_registration).to receive(:enabled).and_return(false)
     end
 
-    describe "GET new" do
+    describe "GET#show" do
       it "redirects to group if signed in" do
         sign_in(person)
 
-        get :new, params: {group_id: group.id}
+        get :show, params: {group_id: group.id}
 
-        is_expected.to redirect_to(group_path(group.id))
+        expect(response).to redirect_to(group_path(group.id))
       end
     end
 
     it "redirects to group if group has registration disabled" do
       group.update!(self_registration_role_type: "")
 
-      get :new, params: {group_id: group.id}
-
-      is_expected.to redirect_to(group_path(group.id))
+      get :show, params: {group_id: group.id}
+      expect(response).to redirect_to(group_path(group.id))
     end
 
     it "redirects to group if group has registration enabled" do
-      get :new, params: {group_id: group.id}
+      get :show, params: {group_id: group.id}
 
-      is_expected.to redirect_to(group_path(group.id))
+      expect(response).to redirect_to(group_path(group.id))
     end
 
     describe "POST create" do
@@ -47,10 +46,10 @@ describe Groups::SelfRegistrationController do
 
         post :create, params: {
           group_id: group.id,
-          main_person_attributes: {email: "foo@example.com"}
+          new_user_form: {email: "foo@example.com"}
         }
 
-        is_expected.to redirect_to(group_path(group.id))
+        expect(response).to redirect_to(group_path(group.id))
       end
     end
   end
@@ -60,7 +59,7 @@ describe Groups::SelfRegistrationController do
       allow(Settings.groups.self_registration).to receive(:enabled).and_return(true)
     end
 
-    context "GET new" do
+    context "GET#show" do
       context "when registration active" do
         before do
           group.update(self_registration_role_type: Group::TopGroup::Member.sti_name)
@@ -68,17 +67,17 @@ describe Groups::SelfRegistrationController do
 
         context "when unautorized" do
           it "renders page" do
-            get :new, params: {group_id: group.id}
+            get :show, params: {group_id: group.id}
 
-            is_expected.to render_template("groups/self_registration/new")
+            expect(response).to render_template("groups/self_registration/show")
           end
 
           it "redirects to group if group has registration disabled" do
             group.update!(self_registration_role_type: "")
 
-            get :new, params: {group_id: group.id}
+            get :show, params: {group_id: group.id}
 
-            is_expected.to redirect_to(group_path(group.id))
+            expect(response).to redirect_to(group_path(group.id))
           end
         end
 
@@ -86,9 +85,9 @@ describe Groups::SelfRegistrationController do
           it "redirects to self_inscription_path" do
             sign_in(person)
 
-            get :new, params: {group_id: group.id}
+            get :show, params: {group_id: group.id}
 
-            is_expected.to redirect_to(group_self_inscription_path(group.id))
+            expect(response).to redirect_to(group_self_inscription_path(group.id))
           end
         end
       end
@@ -97,14 +96,14 @@ describe Groups::SelfRegistrationController do
         it "redirects to group" do
           sign_in(person)
 
-          get :new, params: {group_id: group.id}
+          get :show, params: {group_id: group.id}
 
-          is_expected.to redirect_to(group_path(group.id))
+          expect(response).to redirect_to(group_path(group.id))
         end
       end
     end
 
-    context "POST create" do
+    context "POST#create" do
       context "when registration active" do
         before do
           group.update!(self_registration_role_type: Group::TopGroup::Member.sti_name)
@@ -123,8 +122,8 @@ describe Groups::SelfRegistrationController do
             expect do
               post :create, params: {
                 group_id: group.id,
-                self_registration: {
-                  main_person_attributes: {first_name: "Bob", last_name: "Miller", privacy_policy_accepted: "1"}
+                wizards_register_new_user_wizard: {
+                  new_user_form: {first_name: "Bob", last_name: "Miller", privacy_policy_accepted: "1"}
                 }
               }
             end.to change { Person.count }.by(1)
@@ -140,19 +139,15 @@ describe Groups::SelfRegistrationController do
             expect(role.type).to eq(Group::TopGroup::Member.sti_name)
             expect(role.group).to eq(group)
 
-            is_expected.to redirect_to(new_person_session_path)
+            expect(response).to redirect_to(new_person_session_path)
           end
 
           it "creates person and schedules duplicate location job" do
-            expect(Person::DuplicateLocatorJob).to receive(:new)
-              .with(kind_of(Integer))
-              .and_call_original
-
             expect do
               post :create, params: {
                 group_id: group.id,
-                self_registration: {
-                  main_person_attributes: {first_name: "Alfred", last_name: "Burn", privacy_policy_accepted: "1"}
+                wizards_register_new_user_wizard: {
+                  new_user_form: {first_name: "Alfred", last_name: "Burn", privacy_policy_accepted: "1"}
                 }
               }
             end.to change { Person.count }.by(1)
@@ -167,15 +162,15 @@ describe Groups::SelfRegistrationController do
             person = Person.find_by(first_name: "Alfred", last_name: "Burn")
             expect(person.primary_group).to eq(group)
 
-            is_expected.to redirect_to(new_person_session_path)
+            expect(response).to redirect_to(new_person_session_path)
           end
 
           it "does not create a person if privacy policy is not accepted" do
             expect do
               post :create, params: {
                 group_id: group.id,
-                self_registration: {
-                  main_person_attributes: {first_name: "Bob", last_name: "Miller", privacy_policy_accepted: "0"}
+                wizards_register_new_user_wizard: {
+                  new_user_form: {first_name: "Bob", last_name: "Miller", privacy_policy_accepted: "0"}
                 }
               }
             end.to change { Person.count }.by(0)
@@ -188,18 +183,18 @@ describe Groups::SelfRegistrationController do
           post :create, params: {
             group_id: group.id,
             verification: "foo",
-            main_person_attributes: {email: "foo@example.com"}
+            new_user_form: {email: "foo@example.com"}
           }
 
-          is_expected.to redirect_to(new_person_session_path)
+          expect(response).to redirect_to(new_person_session_path)
         end
 
         it "creates person and role" do
           expect do
             post :create, params: {
               group_id: group.id,
-              self_registration: {
-                main_person_attributes: {first_name: "Bob", last_name: "Miller"}
+              wizards_register_new_user_wizard: {
+                new_user_form: {first_name: "Bob", last_name: "Miller"}
               }
             }
           end.to change { Person.count }.by(1)
@@ -214,7 +209,7 @@ describe Groups::SelfRegistrationController do
           expect(role.type).to eq(Group::TopGroup::Member.sti_name)
           expect(role.group).to eq(group)
 
-          is_expected.to redirect_to(new_person_session_path)
+          expect(response).to redirect_to(new_person_session_path)
         end
 
         it "raises when person save! fails" do
@@ -222,20 +217,8 @@ describe Groups::SelfRegistrationController do
           expect do
             post :create, params: {
               group_id: group.id,
-              self_registration: {
-                main_person_attributes: {first_name: "Bob", last_name: "Miller"}
-              }
-            }
-          end.to raise_error(RuntimeError)
-        end
-
-        it "raises when role save! fails" do
-          allow_any_instance_of(Role).to receive(:save!).and_raise
-          expect do
-            post :create, params: {
-              group_id: group.id,
-              self_registration: {
-                main_person_attributes: {first_name: "Bob", last_name: "Miller"}
+              wizards_register_new_user_wizard: {
+                new_user_form: {first_name: "Bob", last_name: "Miller"}
               }
             }
           end.to raise_error(RuntimeError)
@@ -245,8 +228,8 @@ describe Groups::SelfRegistrationController do
           expect do
             post :create, params: {
               group_id: group.id,
-              self_registration: {
-                main_person_attributes: {first_name: "Bob", last_name: "Miller"}
+              wizards_register_new_user_wizard: {
+                new_user_form: {first_name: "Bob", last_name: "Miller"}
               }
             }
           end.not_to have_enqueued_mail
@@ -256,8 +239,8 @@ describe Groups::SelfRegistrationController do
           expect do
             post :create, params: {
               group_id: group.id,
-              self_registration: {
-                main_person_attributes: {first_name: "Bob", last_name: "Miller", email: "foo@example.com"}
+              wizards_register_new_user_wizard: {
+                new_user_form: {first_name: "Bob", last_name: "Miller", email: "foo@example.com"}
               }
             }
           end.to change { ActionMailer::Base.deliveries.count }.by(1)
@@ -269,8 +252,8 @@ describe Groups::SelfRegistrationController do
           expect do
             post :create, params: {
               group_id: group.id,
-              self_registration: {
-                main_person_attributes: {first_name: "Bob", last_name: "Miller"}
+              wizards_register_new_user_wizard: {
+                new_user_form: {first_name: "Bob", last_name: "Miller"}
               }
             }
           end.to have_enqueued_mail
